2022_11_29
몬스터죽이면 아이템 무제한으로 들어오는 버그 고치십쇼

+++ ProjectA/Assets/03.Ui/Scripts/MainCanvas.cs
@@ -40,9 +40,9 @@ public class MainCanvas : MonoBehaviour
     private void Update()
     {
         //플레이어의 행동력을 Ui형태로 보여줌
-        playerActivePointPer = Player.Instance.nowActivePoint / Player.Instance.maxActivePoint;
+        playerActivePointPer = (float)Player.Instance.nowActivePoint / Player.Instance.maxActivePoint;
         playerActivePoint.transform.GetChild(0).GetComponent<RectTransform>().sizeDelta = new Vector2(0, 150 * playerActivePointPer);
-        playerHpPer = Player.Instance.nowHp / Player.Instance.maxHp;
+        playerHpPer = (float)Player.Instance.nowHp / Player.Instance.maxHp;
         playerHp.transform.GetChild(0).GetComponent<RectTransform>().sizeDelta = new Vector2(0, 150 * playerHpPer);
     }
 }
